Commit because it's been a long time
The new vertical electronics mount has been printed and seems mostly
good, but I plan on making the detents have less stop, and more
positively lock, as well as have them on both sides.

The psu mounts have been brushed up, with the possibility of having the
PSU in the front as well as in the rear, with the option of having the
plug on the left or right rear.

Still left to do:
* rpi camera mounted on X axis motor cage
* rpi mount near bottom of X axis motor area
* display mount for a mini 12864 display